# ClientApp/src/app/API/model/Compte.ts
/**
 * SGC-Miage V1.0
 * JavaInUse API reference for developers
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

'use strict';
import * as models from './models';

export interface Compte {
    

    employeId?: string;

    id?: string;

    mail?: string;

    password?: string;

    typeCompte?: Compte.TypeCompteEnum;
}
export namespace Compte {

    export enum TypeCompteEnum { 
        MEMBRE = <any> 'MEMBRE',
        ADMINISTRATEUR = <any> 'ADMINISTRATEUR',
        MANAGER = <any> 'MANAGER',
    }
}